Organize files and classes in packages
