+++ java/app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true" >
 
         <activity
-            android:name=".CheckoutActivity"
+            android:name=".activity.CheckoutActivity"
             android:label="@string/store_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -20,13 +20,14 @@
         </activity>
 
         <activity
-            android:name=".PaymentTransparentActivity"
+            android:name=".activity.PaymentTransparentActivity"
             android:launchMode="singleTask"
             android:taskAffinity=""
             android:theme="@style/Transparent" >
         </activity>
+
         <service
-            android:name=".PaymentNotificationIntentService"
+            android:name=".service.PaymentNotificationIntentService"
             android:enabled="true"
             android:icon="@drawable/googlepay_button_overlay"
             android:label="@string/app_name" />
