+++ java/app/src/main/res/layout/activity_transparent.xml
@@ -5,5 +5,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".PaymentTransparentActivity">
+    tools:context=".activity.PaymentTransparentActivity">
 </RelativeLayout>
